first version of memory initialisation

+++ src/hash.c
@@ -14,8 +14,8 @@ uint32_t jenkins_hash(const uint8_t* key, size_t length) {
     return hash;
 }
 
-uint8_t node_hash(const uint8_t node_key, const uint8_t node_count) {
-    return node_key % node_count;
+uint32_t node_hash(const uint32_t node_key, const uint8_t node_count) {
+    return (uint32_t) node_key % node_count;
 }
 
 uint32_t local_hash(const uint32_t local_key, const uint32_t local_element_count) {
